feat(79): improve admin layout
