+++ vue/src/domains/admin/components/AdminHeader.vue
@@ -1,119 +1,18 @@
 <script setup lang="ts">
-const { ADMIN_PANEL_HOME, ADMIN_PANEL_ORGANIZATIONS, ADMIN_PANEL_USERS } = routerPageName;
-const route = useRoute();
+import AdminDropdown from "@/domains/admin/components/AdminDropdown.vue";
+import AdminMobileNavbar from "@/domains/admin/components/AdminMobileNavbar.vue";
+
+const user = useUserStore().user;
 </script>
 
 <template>
-	<header class="flex h-24 items-center gap-4 border-b bg-muted/40 px-4 lg:px-6">
-		<TheSheet>
-			<SheetTrigger as-child>
-				<TheButton
-					variant="outline"
-					size="icon"
-					class="shrink-0 md:hidden"
-				>
-					<TheIcon
-						icon="menu"
-						size="2xl"
-					/>
-				</TheButton>
-			</SheetTrigger>
+	<header class="container h-24 flex gap-4 items-center border-b bg-muted/40">
+		<AdminMobileNavbar />
 
-			<SheetContent
-				side="left"
-				class="flex flex-col"
-			>
-				<nav class="grid gap-2 text-lg font-medium">
-					<RouterLink
-						:to="{ name: ADMIN_PANEL_HOME }"
-						class="text-center text-2xl font-bold"
-					>
-						<span>{{ $t("layout.admin.title") }}</span>
-					</RouterLink>
+		<div class="w-full flex-1 text-center md:text-start">
+			<span class=" text-xl">Bonjour {{ user?.firstname }} ! Vous êtes <strong>{{ user?.primordialRole }}</strong>.</span>
+		</div>
 
-					<RouterLink
-						:to="{ name: ADMIN_PANEL_HOME }"
-						class="mx-[-0.65rem] px-3 py-2 flex items-center gap-4 rounded-xl hover:text-foreground"
-						:class="
-							route.name === ADMIN_PANEL_HOME ?
-								'bg-muted text-foreground'
-								:
-								'text-muted-foreground'
-						"
-					>
-						<TheIcon
-							icon="home-outline"
-							size="2xl"
-						/>
-						{{ $t("layout.admin.nav.home") }}
-					</RouterLink>
-
-					<RouterLink
-						:to="{ name: ADMIN_PANEL_ORGANIZATIONS }"
-						class="mx-[-0.65rem] px-3 py-2 flex items-center gap-4 rounded-xl hover:text-foreground"
-						:class="
-							route.name === ADMIN_PANEL_ORGANIZATIONS ?
-								'bg-muted text-foreground'
-								:
-								'text-muted-foreground'
-						"
-					>
-						<TheIcon
-							icon="domain"
-							size="2xl"
-						/>
-						{{ $t("layout.admin.nav.organizations") }}
-					</RouterLink>
-
-					<RouterLink
-						:to="{ name: ADMIN_PANEL_USERS }"
-						class="mx-[-0.65rem] px-3 py-2 flex items-center gap-4 rounded-xl hover:text-foreground"
-						:class="
-							route.name === ADMIN_PANEL_USERS ?
-								'bg-muted text-foreground'
-								:
-								'text-muted-foreground'
-						"
-					>
-						<TheIcon
-							icon="account-multiple"
-							size="2xl"
-						/>
-						{{ $t("layout.admin.nav.users") }}
-					</RouterLink>
-				</nav>
-			</SheetContent>
-		</TheSheet>
-
-		<div class="w-full flex-1" />
-
-		<DropdownMenu>
-			<DropdownMenuTrigger as-child>
-				<TheButton
-					variant="secondary"
-					size="icon"
-					class="rounded-full"
-				>
-					<TheIcon
-						icon="account-circle-outline"
-						size="2xl"
-					/>
-				</TheButton>
-			</DropdownMenuTrigger>
-
-			<DropdownMenuContent align="end">
-				<DropdownMenuLabel>{{ $t("dropdownAccount.myAccount") }}</DropdownMenuLabel>
-
-				<DropdownMenuSeparator />
-
-				<DropdownMenuItem>{{ $t("dropdownAccount.settings") }}</DropdownMenuItem>
-
-				<DropdownMenuItem>{{ $t("dropdownAccount.support") }}</DropdownMenuItem>
-
-				<DropdownMenuSeparator />
-
-				<DropdownMenuItem>{{ $t("dropdownAccount.logout") }}</DropdownMenuItem>
-			</DropdownMenuContent>
-		</DropdownMenu>
+		<AdminDropdown />
 	</header>
 </template>
